requirements.txt
alembic==1.13.1
aniso8601==9.0.1
apturl==0.5.2
asgiref==3.8.1
attrs==23.2.0
bcrypt==3.2.0
black==24.2.0
blinker==1.7.0
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
djangorestframework==3.15.1
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.0
fasteners==0.14.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
future==0.18.2
gpg==1.16.0
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.2
Markdown==3.3.6
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
natsort==8.0.2
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==23.2
paramiko==2.9.3
pathspec==0.12.1
persistent==5.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
pluggy==1.4.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==8.0.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-editor==1.0.4
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.33.0
requests==2.25.1
rpds-py==0.18.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
systemd-python==234
texttable==1.6.4
tomli==2.0.1
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.0.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.2
